arch/arm64: increase speed-bin v3 frequency.
on some Z010D devices the device reads frequencies from speed-bin v3
which by default is 998mhz
these devices were probably manufactured in 2016 but devices which were
manufactured in 2015 reads frequencies from speed-bin v0 and the default
frequency in it is 1209mhz.

This results in lower frequencies in newer devices than the snapdragon
410 provides so add more frequencies steps and make it equal to
speed-bin v0.